@@ -5,6 +5,7 @@
 
 using Microsoft.Health.Core.Features.Context;
 using Microsoft.Health.Fhir.Core.Features.Context;
+using Microsoft.Health.Fhir.Core.Models;
 
 namespace Microsoft.Health.Fhir.Core.Features.Security.Authorization
 {
@@ -21,8 +22,8 @@ namespace Microsoft.Health.Fhir.Core.Features.Security.Authorization
             DataActions permittedDataActions = 0;
             foreach (ScopeRestriction scopeRestriction in allowedResourceActions)
             {
-                // resousourceRequested is null when the base route is queried, i.e. all resources
-                if (scopeRestriction.Resource == "*" || scopeRestriction.Resource == resourceRequested || resourceRequested == null)
+                // resourceRequested is null when the base route is queried, i.e. all resources
+                if (scopeRestriction.Resource == KnownResourceTypes.All || scopeRestriction.Resource == resourceRequested || resourceRequested == null)
                 {
                     permittedDataActions |= scopeRestriction.AllowedDataAction;
                     if (permittedDataActions == dataActions)
