@@ -101,15 +101,13 @@ namespace Microsoft.Health.Fhir.Core.UnitTests.Features.Operations.Import
 
             Progress<string> progress = new Progress<string>();
             ImportProcessingJob job = new ImportProcessingJob(
-                                    inputData,
-                                    result,
                                     loader,
                                     importer,
                                     importErrorStoreFactory,
                                     contextAccessor,
                                     loggerFactory);
 
-            await Assert.ThrowsAsync<RetriableJobException>(() => job.ExecuteAsync(progress, CancellationToken.None));
+            await Assert.ThrowsAsync<RetriableJobException>(() => job.ExecuteAsync(GetJobInfo(inputData, result), progress, CancellationToken.None));
         }
 
         [Fact]
@@ -132,15 +130,13 @@ namespace Microsoft.Health.Fhir.Core.UnitTests.Features.Operations.Import
                 });
 
             ImportProcessingJob job = new ImportProcessingJob(
-                                    inputData,
-                                    result,
                                     loader,
                                     importer,
                                     importErrorStoreFactory,
                                     contextAccessor,
                                     loggerFactory);
 
-            await Assert.ThrowsAsync<JobExecutionException>(() => job.ExecuteAsync(new Progress<string>(), CancellationToken.None));
+            await Assert.ThrowsAsync<JobExecutionException>(() => job.ExecuteAsync(GetJobInfo(inputData, result), new Progress<string>(), CancellationToken.None));
         }
 
         private static async Task VerifyCommonImportAsync(ImportProcessingJobInputData inputData, ImportProcessingJobResult currentResult)
@@ -239,15 +235,13 @@ namespace Microsoft.Health.Fhir.Core.UnitTests.Features.Operations.Import
                 progressResult = r;
             });
             ImportProcessingJob job = new ImportProcessingJob(
-                                    inputData,
-                                    currentResult,
                                     loader,
                                     importer,
                                     importErrorStoreFactory,
                                     contextAccessor,
                                     loggerFactory);
 
-            string resultString = await job.ExecuteAsync(progress, CancellationToken.None);
+            string resultString = await job.ExecuteAsync(GetJobInfo(inputData, currentResult), progress, CancellationToken.None);
             ImportProcessingJobResult result = JsonConvert.DeserializeObject<ImportProcessingJobResult>(resultString);
             Assert.Equal(1 + failedCountFromProgress, result.FailedCount);
             Assert.Equal(1 + succeedCountFromProgress, result.SucceedCount);
@@ -272,6 +266,17 @@ namespace Microsoft.Health.Fhir.Core.UnitTests.Features.Operations.Import
             inputData.UriString = "http://dummy";
 
             return inputData;
+        }
+
+        private static JobInfo GetJobInfo(ImportProcessingJobInputData data, ImportProcessingJobResult result)
+        {
+            var jobInfo = new JobInfo
+            {
+                Definition = JsonConvert.SerializeObject(data),
+                Result = result != null ? JsonConvert.SerializeObject(result) : null,
+            };
+
+            return jobInfo;
         }
     }
 }
