@@ -8,6 +8,7 @@ using System.Collections.Generic;
 using EnsureThat;
 using Microsoft.Health.Core;
 using Microsoft.Health.Fhir.Core.Models;
+using Microsoft.Health.JobManagement;
 using Newtonsoft.Json;
 
 namespace Microsoft.Health.Fhir.Core.Features.Operations.Export.Models
@@ -15,7 +16,7 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Export.Models
     /// <summary>
     /// Class to hold metadata for an individual export request.
     /// </summary>
-    public class ExportJobRecord : JobRecord
+    public class ExportJobRecord : JobRecord, IJobData
     {
         public ExportJobRecord(
             Uri requestUri,
@@ -27,6 +28,7 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Export.Models
             uint rollingFileSizeInMB,
             IReadOnlyCollection<KeyValuePair<string, string>> requestorClaims = null,
             PartialDateTime since = null,
+            PartialDateTime till = null,
             string groupId = null,
             string storageAccountConnectionHash = null,
             string storageAccountUri = null,
@@ -36,7 +38,9 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Export.Models
             uint maximumNumberOfResourcesPerQuery = 100,
             uint numberOfPagesPerCommit = 10,
             string storageAccountContainerName = null,
-            int schemaVersion = 2)
+            int parallel = 0,
+            int schemaVersion = 2,
+            int typeId = (int)JobType.ExportOrchestrator)
         {
             EnsureArg.IsNotNull(requestUri, nameof(requestUri));
             EnsureArg.IsNotNullOrWhiteSpace(hash, nameof(hash));
@@ -58,6 +62,8 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Export.Models
             NumberOfPagesPerCommit = numberOfPagesPerCommit;
             RollingFileSizeInMB = rollingFileSizeInMB;
             RestartCount = 0;
+            TypeId = typeId;
+            Parallel = parallel;
 
             AnonymizationConfigurationCollectionReference = anonymizationConfigurationCollectionReference;
             AnonymizationConfigurationLocation = anonymizationConfigurationLocation;
@@ -69,7 +75,7 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Export.Models
             Status = OperationStatus.Queued;
 
             QueuedTime = Clock.UtcNow;
-            Till = new PartialDateTime(Clock.UtcNow);
+            Till = till ?? new PartialDateTime(Clock.UtcNow);
 
             if (string.IsNullOrWhiteSpace(storageAccountContainerName))
             {
@@ -85,6 +91,9 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Export.Models
         protected ExportJobRecord()
         {
         }
+
+        [JsonProperty(JobRecordProperties.TypeId)]
+        public int TypeId { get; internal set; }
 
         [JsonProperty(JobRecordProperties.RequestUri)]
         public Uri RequestUri { get; internal set; }
@@ -164,6 +173,9 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Export.Models
 
         [JsonProperty(JobRecordProperties.RestartCount)]
         public uint RestartCount { get; set; }
+
+        [JsonProperty(JobRecordProperties.Parallel)]
+        public int Parallel { get; private set; }
 
         internal ExportJobRecord Clone()
         {
