@@ -12,6 +12,7 @@ using System.Threading.Tasks;
 using EnsureThat;
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using Microsoft.Extensions.Primitives;
 using Microsoft.Health.Core;
 using Microsoft.Health.Core.Features.Context;
@@ -23,71 +24,59 @@ using JobStatus = Microsoft.Health.JobManagement.JobStatus;
 
 namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
 {
+    [JobTypeId((int)JobType.ImportOrchestrator)]
     public class ImportOrchestratorJob : IJob
     {
-        public const short ImportOrchestratorTypeId = 2;
-
         private const long DefaultResourceSizePerByte = 64;
 
         private readonly IMediator _mediator;
-        private ImportOrchestratorJobInputData _orchestratorInputData;
-        private RequestContextAccessor<IFhirRequestContext> _contextAccessor;
-        private ImportOrchestratorJobResult _orchestratorJobResult;
-        private IImportOrchestratorJobDataStoreOperation _importOrchestratorJobDataStoreOperation;
-        private IQueueClient _queueClient;
-        private JobInfo _jobInfo;
+        private readonly RequestContextAccessor<IFhirRequestContext> _contextAccessor;
+        private readonly IImportOrchestratorJobDataStoreOperation _importOrchestratorJobDataStoreOperation;
+        private readonly IQueueClient _queueClient;
         private ImportTaskConfiguration _importConfiguration;
         private ILogger<ImportOrchestratorJob> _logger;
         private IIntegrationDataStoreClient _integrationDataStoreClient;
 
         public ImportOrchestratorJob(
             IMediator mediator,
-            ImportOrchestratorJobInputData orchestratorInputData,
-            ImportOrchestratorJobResult orchrestratorJobResult,
             RequestContextAccessor<IFhirRequestContext> contextAccessor,
             IImportOrchestratorJobDataStoreOperation importOrchestratorJobDataStoreOperation,
             IIntegrationDataStoreClient integrationDataStoreClient,
             IQueueClient queueClient,
-            JobInfo jobInfo,
-            ImportTaskConfiguration importConfiguration,
+            IOptions<ImportTaskConfiguration> importConfiguration,
             ILoggerFactory loggerFactory)
         {
             EnsureArg.IsNotNull(mediator, nameof(mediator));
-            EnsureArg.IsNotNull(orchestratorInputData, nameof(orchestratorInputData));
-            EnsureArg.IsNotNull(orchrestratorJobResult, nameof(orchrestratorJobResult));
             EnsureArg.IsNotNull(contextAccessor, nameof(contextAccessor));
             EnsureArg.IsNotNull(importOrchestratorJobDataStoreOperation, nameof(importOrchestratorJobDataStoreOperation));
             EnsureArg.IsNotNull(integrationDataStoreClient, nameof(integrationDataStoreClient));
             EnsureArg.IsNotNull(queueClient, nameof(queueClient));
-            EnsureArg.IsNotNull(jobInfo, nameof(jobInfo));
-            EnsureArg.IsNotNull(importConfiguration, nameof(importConfiguration));
+            EnsureArg.IsNotNull(importConfiguration?.Value, nameof(importConfiguration));
             EnsureArg.IsNotNull(loggerFactory, nameof(loggerFactory));
 
             _mediator = mediator;
-            _orchestratorInputData = orchestratorInputData;
-            _orchestratorJobResult = orchrestratorJobResult;
             _contextAccessor = contextAccessor;
             _importOrchestratorJobDataStoreOperation = importOrchestratorJobDataStoreOperation;
             _integrationDataStoreClient = integrationDataStoreClient;
             _queueClient = queueClient;
-            _jobInfo = jobInfo;
-            _importConfiguration = importConfiguration;
+            _importConfiguration = importConfiguration.Value;
             _logger = loggerFactory.CreateLogger<ImportOrchestratorJob>();
 
             PollingFrequencyInSeconds = _importConfiguration.PollingFrequencyInSeconds;
         }
 
-        public string RunId { get; set; }
-
         public int PollingFrequencyInSeconds { get; set; }
 
-        public async Task<string> ExecuteAsync(IProgress<string> progress, CancellationToken cancellationToken)
+        public async Task<string> ExecuteAsync(JobInfo jobInfo, IProgress<string> progress, CancellationToken cancellationToken)
         {
+            ImportOrchestratorJobInputData inputData = JsonConvert.DeserializeObject<ImportOrchestratorJobInputData>(jobInfo.Definition);
+            ImportOrchestratorJobResult currentResult = string.IsNullOrEmpty(jobInfo.Result) ? new ImportOrchestratorJobResult() : JsonConvert.DeserializeObject<ImportOrchestratorJobResult>(jobInfo.Result);
+
             var fhirRequestContext = new FhirRequestContext(
                     method: "Import",
-                    uriString: _orchestratorInputData.RequestUri.ToString(),
-                    baseUriString: _orchestratorInputData.BaseUri.ToString(),
-                    correlationId: _jobInfo.Id.ToString(),
+                    uriString: inputData.RequestUri.ToString(),
+                    baseUriString: inputData.BaseUri.ToString(),
+                    correlationId: jobInfo.Id.ToString(),
                     requestHeaders: new Dictionary<string, StringValues>(),
                     responseHeaders: new Dictionary<string, StringValues>())
             {
@@ -96,8 +85,8 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
 
             _contextAccessor.RequestContext = fhirRequestContext;
 
-            _orchestratorJobResult.Request = _orchestratorInputData.RequestUri.ToString();
-            _orchestratorJobResult.TransactionTime = _orchestratorInputData.CreateTime;
+            currentResult.Request = inputData.RequestUri.ToString();
+            currentResult.TransactionTime = inputData.CreateTime;
 
             ImportOrchestratorJobErrorResult errorResult = null;
 
@@ -108,32 +97,32 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
                     throw new OperationCanceledException();
                 }
 
-                if (_orchestratorJobResult.Progress == ImportOrchestratorJobProgress.Initialized)
+                if (currentResult.Progress == ImportOrchestratorJobProgress.Initialized)
                 {
-                    await ValidateResourcesAsync(cancellationToken);
+                    await ValidateResourcesAsync(inputData, cancellationToken);
 
-                    _orchestratorJobResult.Progress = ImportOrchestratorJobProgress.InputResourcesValidated;
-                    progress.Report(JsonConvert.SerializeObject(_orchestratorJobResult));
+                    currentResult.Progress = ImportOrchestratorJobProgress.InputResourcesValidated;
+                    progress.Report(JsonConvert.SerializeObject(currentResult));
 
                     _logger.LogInformation("Input Resources Validated");
                 }
 
-                if (_orchestratorJobResult.Progress == ImportOrchestratorJobProgress.InputResourcesValidated)
+                if (currentResult.Progress == ImportOrchestratorJobProgress.InputResourcesValidated)
                 {
                     await _importOrchestratorJobDataStoreOperation.PreprocessAsync(cancellationToken);
 
-                    _orchestratorJobResult.Progress = ImportOrchestratorJobProgress.PreprocessCompleted;
-                    _orchestratorJobResult.CurrentSequenceId = _orchestratorInputData.StartSequenceId;
-                    progress.Report(JsonConvert.SerializeObject(_orchestratorJobResult));
+                    currentResult.Progress = ImportOrchestratorJobProgress.PreprocessCompleted;
+                    currentResult.CurrentSequenceId = inputData.StartSequenceId;
+                    progress.Report(JsonConvert.SerializeObject(currentResult));
 
                     _logger.LogInformation("Preprocess Completed");
                 }
 
-                if (_orchestratorJobResult.Progress == ImportOrchestratorJobProgress.PreprocessCompleted)
+                if (currentResult.Progress == ImportOrchestratorJobProgress.PreprocessCompleted)
                 {
-                    await ExecuteImportProcessingJobAsync(progress, cancellationToken);
-                    _orchestratorJobResult.Progress = ImportOrchestratorJobProgress.SubJobsCompleted;
-                    progress.Report(JsonConvert.SerializeObject(_orchestratorJobResult));
+                    await ExecuteImportProcessingJobAsync(progress, jobInfo, inputData, currentResult, cancellationToken);
+                    currentResult.Progress = ImportOrchestratorJobProgress.SubJobsCompleted;
+                    progress.Report(JsonConvert.SerializeObject(currentResult));
 
                     _logger.LogInformation("SubJobs Completed");
                 }
@@ -149,8 +138,8 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
                 };
 
                 // Processing jobs has been cancelled by CancelImportRequestHandler
-                await WaitCancelledJobCompletedAsync();
-                await SendImportMetricsNotification(JobStatus.Cancelled);
+                await WaitCancelledJobCompletedAsync(jobInfo);
+                await SendImportMetricsNotification(JobStatus.Cancelled, jobInfo, inputData, currentResult);
             }
             catch (OperationCanceledException canceledEx)
             {
@@ -163,8 +152,8 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
                 };
 
                 // Processing jobs has been cancelled by CancelImportRequestHandler
-                await WaitCancelledJobCompletedAsync();
-                await SendImportMetricsNotification(JobStatus.Cancelled);
+                await WaitCancelledJobCompletedAsync(jobInfo);
+                await SendImportMetricsNotification(JobStatus.Cancelled, jobInfo, inputData, currentResult);
             }
             catch (IntegrationDataStoreException integrationDataStoreEx)
             {
@@ -176,7 +165,7 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
                     ErrorMessage = integrationDataStoreEx.Message,
                 };
 
-                await SendImportMetricsNotification(JobStatus.Failed);
+                await SendImportMetricsNotification(JobStatus.Failed, jobInfo, inputData, currentResult);
             }
             catch (ImportFileEtagNotMatchException eTagEx)
             {
@@ -188,7 +177,7 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
                     ErrorMessage = eTagEx.Message,
                 };
 
-                await SendImportMetricsNotification(JobStatus.Failed);
+                await SendImportMetricsNotification(JobStatus.Failed, jobInfo, inputData, currentResult);
             }
             catch (ImportProcessingException processingEx)
             {
@@ -201,8 +190,8 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
                 };
 
                 // Cancel other processing jobs
-                await CancelProcessingJobsAsync();
-                await SendImportMetricsNotification(JobStatus.Failed);
+                await CancelProcessingJobsAsync(jobInfo);
+                await SendImportMetricsNotification(JobStatus.Failed, jobInfo, inputData, currentResult);
             }
             catch (RetriableJobException ex)
             {
@@ -221,8 +210,8 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
                 };
 
                 // Cancel processing jobs for critical error in orchestrator job
-                await CancelProcessingJobsAsync();
-                await SendImportMetricsNotification(JobStatus.Failed);
+                await CancelProcessingJobsAsync(jobInfo);
+                await SendImportMetricsNotification(JobStatus.Failed, jobInfo, inputData, currentResult);
             }
 
             // Post-process operation cannot be cancelled.
@@ -250,8 +239,8 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
                 throw new JobExecutionException(errorResult.ErrorMessage, errorResult);
             }
 
-            await SendImportMetricsNotification(JobStatus.Completed);
-            return JsonConvert.SerializeObject(_orchestratorJobResult);
+            await SendImportMetricsNotification(JobStatus.Completed, jobInfo, inputData, currentResult);
+            return JsonConvert.SerializeObject(currentResult);
         }
 
         private static long CalculateResourceNumberByResourceSize(long blobSizeInBytes, long resourceCountPerBytes)
@@ -259,9 +248,9 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
             return Math.Max((blobSizeInBytes / resourceCountPerBytes) + 1, 10000L);
         }
 
-        private async Task ValidateResourcesAsync(CancellationToken cancellationToken)
+        private async Task ValidateResourcesAsync(ImportOrchestratorJobInputData inputData, CancellationToken cancellationToken)
         {
-            foreach (var input in _orchestratorInputData.Input)
+            foreach (var input in inputData.Input)
             {
                 Dictionary<string, object> properties = await _integrationDataStoreClient.GetPropertiesAsync(input.Url, cancellationToken);
                 if (!string.IsNullOrEmpty(input.Etag))
@@ -274,58 +263,58 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
             }
         }
 
-        private async Task SendImportMetricsNotification(JobStatus jobStatus)
+        private async Task SendImportMetricsNotification(JobStatus jobStatus, JobInfo jobInfo, ImportOrchestratorJobInputData inputData, ImportOrchestratorJobResult currentResult)
         {
             ImportJobMetricsNotification importJobMetricsNotification = new ImportJobMetricsNotification(
-                _jobInfo.Id.ToString(),
+                jobInfo.Id.ToString(),
                 jobStatus.ToString(),
-                _orchestratorInputData.CreateTime,
+                inputData.CreateTime,
                 Clock.UtcNow,
-                _orchestratorJobResult.TotalSizeInBytes,
-                _orchestratorJobResult.SucceedImportCount,
-                _orchestratorJobResult.FailedImportCount);
+                currentResult.TotalSizeInBytes,
+                currentResult.SucceedImportCount,
+                currentResult.FailedImportCount);
 
             await _mediator.Publish(importJobMetricsNotification, CancellationToken.None);
         }
 
-        private async Task ExecuteImportProcessingJobAsync(IProgress<string> progress, CancellationToken cancellationToken)
+        private async Task ExecuteImportProcessingJobAsync(IProgress<string> progress, JobInfo jobInfo, ImportOrchestratorJobInputData inputData, ImportOrchestratorJobResult currentResult, CancellationToken cancellationToken)
         {
-            _orchestratorJobResult.TotalSizeInBytes = _orchestratorJobResult.TotalSizeInBytes ?? 0;
+            currentResult.TotalSizeInBytes = currentResult.TotalSizeInBytes ?? 0;
 
-            foreach (var input in _orchestratorInputData.Input.Skip(_orchestratorJobResult.CreatedJobCount))
+            foreach (var input in inputData.Input.Skip(currentResult.CreatedJobCount))
             {
                 if (cancellationToken.IsCancellationRequested)
                 {
                     throw new OperationCanceledException();
                 }
 
-                while (_orchestratorJobResult.RunningJobIds.Count >= _importConfiguration.MaxRunningProcessingJobCount)
+                while (currentResult.RunningJobIds.Count >= _importConfiguration.MaxRunningProcessingJobCount)
                 {
-                    await WaitRunningJobComplete(progress, cancellationToken);
+                    await WaitRunningJobComplete(progress, jobInfo, currentResult, cancellationToken);
                 }
 
-                (long processingJobId, long endSequenceId, long blobSizeInBytes) = await CreateNewProcessingJobAsync(input, cancellationToken);
+                (long processingJobId, long endSequenceId, long blobSizeInBytes) = await CreateNewProcessingJobAsync(input, jobInfo, inputData, currentResult, cancellationToken);
 
-                _orchestratorJobResult.RunningJobIds.Add(processingJobId);
-                _orchestratorJobResult.CurrentSequenceId = endSequenceId;
-                _orchestratorJobResult.TotalSizeInBytes += blobSizeInBytes;
-                _orchestratorJobResult.CreatedJobCount += 1;
-                progress.Report(JsonConvert.SerializeObject(_orchestratorJobResult));
+                currentResult.RunningJobIds.Add(processingJobId);
+                currentResult.CurrentSequenceId = endSequenceId;
+                currentResult.TotalSizeInBytes += blobSizeInBytes;
+                currentResult.CreatedJobCount += 1;
+                progress.Report(JsonConvert.SerializeObject(currentResult));
             }
 
-            while (_orchestratorJobResult.RunningJobIds.Count > 0)
+            while (currentResult.RunningJobIds.Count > 0)
             {
-                await WaitRunningJobComplete(progress, cancellationToken);
+                await WaitRunningJobComplete(progress, jobInfo, currentResult, cancellationToken);
             }
         }
 
-        private async Task WaitRunningJobComplete(IProgress<string> progress, CancellationToken cancellationToken)
+        private async Task WaitRunningJobComplete(IProgress<string> progress, JobInfo jobInfo, ImportOrchestratorJobResult currentResult, CancellationToken cancellationToken)
         {
             HashSet<long> completedJobIds = new HashSet<long>();
             List<JobInfo> runningJobs = new List<JobInfo>();
             try
             {
-                runningJobs.AddRange(await _queueClient.GetJobsByIdsAsync(_jobInfo.QueueType, _orchestratorJobResult.RunningJobIds.ToArray(), false, cancellationToken));
+                runningJobs.AddRange(await _queueClient.GetJobsByIdsAsync(jobInfo.QueueType, currentResult.RunningJobIds.ToArray(), false, cancellationToken));
             }
             catch (Exception ex)
             {
@@ -345,8 +334,8 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
                     if (latestJobInfo.Status == JobStatus.Completed)
                     {
                         ImportProcessingJobResult procesingJobResult = JsonConvert.DeserializeObject<ImportProcessingJobResult>(latestJobInfo.Result);
-                        _orchestratorJobResult.SucceedImportCount += procesingJobResult.SucceedCount;
-                        _orchestratorJobResult.FailedImportCount += procesingJobResult.FailedCount;
+                        currentResult.SucceedImportCount += procesingJobResult.SucceedCount;
+                        currentResult.FailedImportCount += procesingJobResult.FailedCount;
                     }
                     else if (latestJobInfo.Status == JobStatus.Failed)
                     {
@@ -364,8 +353,8 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
 
             if (completedJobIds.Count > 0)
             {
-                _orchestratorJobResult.RunningJobIds.ExceptWith(completedJobIds);
-                progress.Report(JsonConvert.SerializeObject(_orchestratorJobResult));
+                currentResult.RunningJobIds.ExceptWith(completedJobIds);
+                progress.Report(JsonConvert.SerializeObject(currentResult));
             }
             else
             {
@@ -374,31 +363,31 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
             }
         }
 
-        private async Task<(long jobId, long endSequenceId, long blobSizeInBytes)> CreateNewProcessingJobAsync(Models.InputResource input, CancellationToken cancellationToken)
+        private async Task<(long jobId, long endSequenceId, long blobSizeInBytes)> CreateNewProcessingJobAsync(Models.InputResource input, JobInfo jobInfo, ImportOrchestratorJobInputData inputData, ImportOrchestratorJobResult currentResult, CancellationToken cancellationToken)
         {
             Dictionary<string, object> properties = await _integrationDataStoreClient.GetPropertiesAsync(input.Url, cancellationToken);
             long blobSizeInBytes = (long)properties[IntegrationDataStoreClientConstants.BlobPropertyLength];
             long estimatedResourceNumber = CalculateResourceNumberByResourceSize(blobSizeInBytes, DefaultResourceSizePerByte);
-            long beginSequenceId = _orchestratorJobResult.CurrentSequenceId;
+            long beginSequenceId = currentResult.CurrentSequenceId;
             long endSequenceId = beginSequenceId + estimatedResourceNumber;
 
             ImportProcessingJobInputData importJobPayload = new ImportProcessingJobInputData()
             {
-                TypeId = ImportProcessingJob.ImportProcessingJobTypeId,
+                TypeId = (int)JobType.ImportProcessing,
                 ResourceLocation = input.Url.ToString(),
-                UriString = _orchestratorInputData.RequestUri.ToString(),
-                BaseUriString = _orchestratorInputData.BaseUri.ToString(),
+                UriString = inputData.RequestUri.ToString(),
+                BaseUriString = inputData.BaseUri.ToString(),
                 ResourceType = input.Type,
                 BeginSequenceId = beginSequenceId,
                 EndSequenceId = endSequenceId,
-                JobId = $"{_jobInfo.GroupId}_{beginSequenceId}",
+                JobId = $"{jobInfo.GroupId}_{beginSequenceId}",
             };
 
             string[] definitions = new string[] { JsonConvert.SerializeObject(importJobPayload) };
 
             try
             {
-                JobInfo jobInfoFromServer = (await _queueClient.EnqueueAsync(_jobInfo.QueueType, definitions, _jobInfo.GroupId, false, false, cancellationToken)).First();
+                JobInfo jobInfoFromServer = (await _queueClient.EnqueueAsync(jobInfo.QueueType, definitions, jobInfo.GroupId, false, false, cancellationToken)).First();
 
                 return (jobInfoFromServer.Id, endSequenceId, blobSizeInBytes);
             }
@@ -409,36 +398,36 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
             }
         }
 
-        private async Task CancelProcessingJobsAsync()
+        private async Task CancelProcessingJobsAsync(JobInfo jobInfo)
         {
             try
             {
-                await _queueClient.CancelJobByGroupIdAsync(_jobInfo.QueueType, _jobInfo.GroupId, CancellationToken.None);
+                await _queueClient.CancelJobByGroupIdAsync(jobInfo.QueueType, jobInfo.GroupId, CancellationToken.None);
             }
             catch (Exception ex)
             {
-                _logger.LogWarning(ex, "failed to cancel job {GroupId}", _jobInfo.GroupId);
+                _logger.LogWarning(ex, "failed to cancel job {GroupId}", jobInfo.GroupId);
             }
 
-            await WaitCancelledJobCompletedAsync();
+            await WaitCancelledJobCompletedAsync(jobInfo);
         }
 
-        private async Task WaitCancelledJobCompletedAsync()
+        private async Task WaitCancelledJobCompletedAsync(JobInfo jobInfo)
         {
             while (true)
             {
                 try
                 {
-                    IEnumerable<JobInfo> jobInfos = await _queueClient.GetJobByGroupIdAsync((byte)QueueType.Import, _jobInfo.GroupId, false, CancellationToken.None);
+                    IEnumerable<JobInfo> jobInfos = await _queueClient.GetJobByGroupIdAsync((byte)QueueType.Import, jobInfo.GroupId, false, CancellationToken.None);
 
-                    if (jobInfos.All(t => (t.Status != JobStatus.Created && t.Status != JobStatus.Running) || !t.CancelRequested || t.Id == _jobInfo.Id))
+                    if (jobInfos.All(t => (t.Status != JobStatus.Created && t.Status != JobStatus.Running) || !t.CancelRequested || t.Id == jobInfo.Id))
                     {
                         break;
                     }
                 }
                 catch (Exception ex)
                 {
-                    _logger.LogWarning(ex, "failed to get jobs by groupId {GroupId}", _jobInfo.GroupId);
+                    _logger.LogWarning(ex, "failed to get jobs by groupId {GroupId}", jobInfo.GroupId);
                     throw new RetriableJobException(ex.Message, ex);
                 }
 
