@@ -18,38 +18,30 @@ using Newtonsoft.Json;
 
 namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
 {
+    [JobTypeId((int)JobType.ImportProcessing)]
     public class ImportProcessingJob : IJob
     {
-        public const short ImportProcessingJobTypeId = 1;
-        public const string CancelledErrorMessage = "Data processing job is canceled.";
+        private const string CancelledErrorMessage = "Data processing job is canceled.";
 
-        private ImportProcessingJobInputData _inputData;
-        private ImportProcessingJobResult _importResult;
-        private IImportResourceLoader _importResourceLoader;
-        private IResourceBulkImporter _resourceBulkImporter;
-        private IImportErrorStoreFactory _importErrorStoreFactory;
-        private RequestContextAccessor<IFhirRequestContext> _contextAccessor;
-        private ILogger<ImportProcessingJob> _logger;
+        private readonly IImportResourceLoader _importResourceLoader;
+        private readonly IResourceBulkImporter _resourceBulkImporter;
+        private readonly IImportErrorStoreFactory _importErrorStoreFactory;
+        private readonly RequestContextAccessor<IFhirRequestContext> _contextAccessor;
+        private readonly ILogger<ImportProcessingJob> _logger;
 
         public ImportProcessingJob(
-            ImportProcessingJobInputData inputData,
-            ImportProcessingJobResult importResult,
             IImportResourceLoader importResourceLoader,
             IResourceBulkImporter resourceBulkImporter,
             IImportErrorStoreFactory importErrorStoreFactory,
             RequestContextAccessor<IFhirRequestContext> contextAccessor,
             ILoggerFactory loggerFactory)
         {
-            EnsureArg.IsNotNull(inputData, nameof(inputData));
-            EnsureArg.IsNotNull(importResult, nameof(importResult));
             EnsureArg.IsNotNull(importResourceLoader, nameof(importResourceLoader));
             EnsureArg.IsNotNull(resourceBulkImporter, nameof(resourceBulkImporter));
             EnsureArg.IsNotNull(importErrorStoreFactory, nameof(importErrorStoreFactory));
             EnsureArg.IsNotNull(contextAccessor, nameof(contextAccessor));
             EnsureArg.IsNotNull(loggerFactory, nameof(loggerFactory));
 
-            _inputData = inputData;
-            _importResult = importResult;
             _importResourceLoader = importResourceLoader;
             _resourceBulkImporter = resourceBulkImporter;
             _importErrorStoreFactory = importErrorStoreFactory;
@@ -58,15 +50,19 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
             _logger = loggerFactory.CreateLogger<ImportProcessingJob>();
         }
 
-        public string RunId { get; set; }
-
-        public async Task<string> ExecuteAsync(IProgress<string> progress, CancellationToken cancellationToken)
+        public async Task<string> ExecuteAsync(JobInfo jobInfo, IProgress<string> progress, CancellationToken cancellationToken)
         {
+            EnsureArg.IsNotNull(jobInfo, nameof(jobInfo));
+            EnsureArg.IsNotNull(progress, nameof(progress));
+
+            ImportProcessingJobInputData inputData = JsonConvert.DeserializeObject<ImportProcessingJobInputData>(jobInfo.Definition);
+            ImportProcessingJobResult currentResult = string.IsNullOrEmpty(jobInfo.Result) ? new ImportProcessingJobResult() : JsonConvert.DeserializeObject<ImportProcessingJobResult>(jobInfo.Result);
+
             var fhirRequestContext = new FhirRequestContext(
                     method: "Import",
-                    uriString: _inputData.UriString,
-                    baseUriString: _inputData.BaseUriString,
-                    correlationId: _inputData.JobId,
+                    uriString: inputData.UriString,
+                    baseUriString: inputData.BaseUriString,
+                    correlationId: inputData.JobId,
                     requestHeaders: new Dictionary<string, StringValues>(),
                     responseHeaders: new Dictionary<string, StringValues>())
             {
@@ -75,12 +71,12 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
 
             _contextAccessor.RequestContext = fhirRequestContext;
 
-            long succeedImportCount = _importResult.SucceedCount;
-            long failedImportCount = _importResult.FailedCount;
+            long succeedImportCount = currentResult.SucceedCount;
+            long failedImportCount = currentResult.FailedCount;
 
-            _importResult.ResourceType = _inputData.ResourceType;
-            _importResult.ResourceLocation = _inputData.ResourceLocation;
-            progress.Report(JsonConvert.SerializeObject(_importResult));
+            currentResult.ResourceType = inputData.ResourceType;
+            currentResult.ResourceLocation = inputData.ResourceLocation;
+            progress.Report(JsonConvert.SerializeObject(currentResult));
 
             try
             {
@@ -89,17 +85,17 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
                     throw new OperationCanceledException();
                 }
 
-                Func<long, long> sequenceIdGenerator = (index) => _inputData.BeginSequenceId + index;
+                Func<long, long> sequenceIdGenerator = (index) => inputData.BeginSequenceId + index;
 
                 // Clean resources before import start
-                await _resourceBulkImporter.CleanResourceAsync(_inputData, _importResult, cancellationToken);
+                await _resourceBulkImporter.CleanResourceAsync(inputData, currentResult, cancellationToken);
 
                 // Initialize error store
-                IImportErrorStore importErrorStore = await _importErrorStoreFactory.InitializeAsync(GetErrorFileName(), cancellationToken);
-                _importResult.ErrorLogLocation = importErrorStore.ErrorFileLocation;
+                IImportErrorStore importErrorStore = await _importErrorStoreFactory.InitializeAsync(GetErrorFileName(inputData), cancellationToken);
+                currentResult.ErrorLogLocation = importErrorStore.ErrorFileLocation;
 
                 // Load and parse resource from bulk resource
-                (Channel<ImportResource> importResourceChannel, Task loadTask) = _importResourceLoader.LoadResources(_inputData.ResourceLocation, _importResult.CurrentIndex, _inputData.ResourceType, sequenceIdGenerator, cancellationToken);
+                (Channel<ImportResource> importResourceChannel, Task loadTask) = _importResourceLoader.LoadResources(inputData.ResourceLocation, currentResult.CurrentIndex, inputData.ResourceType, sequenceIdGenerator, cancellationToken);
 
                 // Import to data store
                 (Channel<ImportProcessingProgress> progressChannel, Task importTask) = _resourceBulkImporter.Import(importResourceChannel, importErrorStore, cancellationToken);
@@ -112,12 +108,12 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
                         throw new OperationCanceledException("Import job is canceled by user.");
                     }
 
-                    _importResult.SucceedCount = batchProgress.SucceedImportCount + succeedImportCount;
-                    _importResult.FailedCount = batchProgress.FailedImportCount + failedImportCount;
-                    _importResult.CurrentIndex = batchProgress.CurrentIndex;
+                    currentResult.SucceedCount = batchProgress.SucceedImportCount + succeedImportCount;
+                    currentResult.FailedCount = batchProgress.FailedImportCount + failedImportCount;
+                    currentResult.CurrentIndex = batchProgress.CurrentIndex;
 
-                    _logger.LogInformation("Import job progress: succeed {SucceedCount}, failed: {FailedCount}", _importResult.SucceedCount, _importResult.FailedCount);
-                    progress.Report(JsonConvert.SerializeObject(_importResult));
+                    _logger.LogInformation("Import job progress: succeed {SucceedCount}, failed: {FailedCount}", currentResult.SucceedCount, currentResult.FailedCount);
+                    progress.Report(JsonConvert.SerializeObject(currentResult));
                 }
 
                 // Pop up exception during load & import
@@ -150,13 +146,13 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
                     throw new RetriableJobException("Failed to load data", ex);
                 }
 
-                return JsonConvert.SerializeObject(_importResult);
+                return JsonConvert.SerializeObject(currentResult);
             }
             catch (TaskCanceledException canceledEx)
             {
                 _logger.LogInformation(canceledEx, CancelledErrorMessage);
 
-                await CleanResourceForFailureAsync();
+                await CleanResourceForFailureAsync(inputData, currentResult);
 
                 ImportProcessingJobErrorResult error = new ImportProcessingJobErrorResult()
                 {
@@ -169,7 +165,7 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
             {
                 _logger.LogInformation(canceledEx, "Data processing task is canceled.");
 
-                await CleanResourceForFailureAsync();
+                await CleanResourceForFailureAsync(inputData, currentResult);
 
                 ImportProcessingJobErrorResult error = new ImportProcessingJobErrorResult()
                 {
@@ -182,7 +178,7 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
             {
                 _logger.LogInformation(retriableEx, "Error in data processing job.");
 
-                await CleanResourceForFailureAsync();
+                await CleanResourceForFailureAsync(inputData, currentResult);
 
                 throw;
             }
@@ -190,7 +186,7 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
             {
                 _logger.LogInformation(ex, "Critical error in data processing job.");
 
-                await CleanResourceForFailureAsync();
+                await CleanResourceForFailureAsync(inputData, currentResult);
 
                 ImportProcessingJobErrorResult error = new ImportProcessingJobErrorResult()
                 {
@@ -204,11 +200,11 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
         /// <summary>
         /// Try best to clean failure data.
         /// </summary>
-        private async Task CleanResourceForFailureAsync()
+        private async Task CleanResourceForFailureAsync(ImportProcessingJobInputData inputData, ImportProcessingJobResult currentResult)
         {
             try
             {
-                await _resourceBulkImporter.CleanResourceAsync(_inputData, _importResult, CancellationToken.None);
+                await _resourceBulkImporter.CleanResourceAsync(inputData, currentResult, CancellationToken.None);
             }
             catch (Exception ex)
             {
@@ -216,9 +212,9 @@ namespace Microsoft.Health.Fhir.Core.Features.Operations.Import
             }
         }
 
-        private string GetErrorFileName()
+        private static string GetErrorFileName(ImportProcessingJobInputData inputData)
         {
-            return $"{_inputData.ResourceType}{_inputData.JobId}.ndjson";
+            return $"{inputData.ResourceType}{inputData.JobId}.ndjson";
         }
     }
 }
