@@ -9,6 +9,8 @@ using System.Globalization;
 using System.Linq;
 using EnsureThat;
 using Hl7.Fhir.Utility;
+using Microsoft.Health.Core.Features.Context;
+using Microsoft.Health.Fhir.Core.Features.Context;
 using Microsoft.Health.Fhir.Core.Features.Definition;
 using Microsoft.Health.Fhir.Core.Features.Persistence;
 using Microsoft.Health.Fhir.Core.Models;
@@ -30,6 +32,7 @@ namespace Microsoft.Health.Fhir.Core.Features.Search.Expressions.Parsers
 
         private readonly ISearchParameterDefinitionManager _searchParameterDefinitionManager;
         private readonly ISearchParameterExpressionParser _searchParameterExpressionParser;
+        private readonly RequestContextAccessor<IFhirRequestContext> _contextAccessor;
 
         private const char SearchSplitChar = ':';
         private const char ChainParameter = '.';
@@ -40,15 +43,19 @@ namespace Microsoft.Health.Fhir.Core.Features.Search.Expressions.Parsers
         /// </summary>
         /// <param name="searchParameterDefinitionManagerResolver">The search parameter definition manager.</param>
         /// <param name="searchParameterExpressionParser">The parser used to parse the search value into a search expression.</param>
+        /// /// <param name="contextAccessor">Used to access the FhirRequestContext</param>
         public ExpressionParser(
             ISearchParameterDefinitionManager.SearchableSearchParameterDefinitionManagerResolver searchParameterDefinitionManagerResolver,
-            ISearchParameterExpressionParser searchParameterExpressionParser)
+            ISearchParameterExpressionParser searchParameterExpressionParser,
+            RequestContextAccessor<IFhirRequestContext> contextAccessor)
         {
             EnsureArg.IsNotNull(searchParameterDefinitionManagerResolver, nameof(searchParameterDefinitionManagerResolver));
             EnsureArg.IsNotNull(searchParameterExpressionParser, nameof(searchParameterExpressionParser));
+            EnsureArg.IsNotNull(contextAccessor, nameof(contextAccessor));
 
             _searchParameterDefinitionManager = searchParameterDefinitionManagerResolver();
             _searchParameterExpressionParser = searchParameterExpressionParser;
+            _contextAccessor = contextAccessor;
         }
 
         /// <summary>
@@ -219,6 +226,15 @@ namespace Microsoft.Health.Fhir.Core.Features.Search.Expressions.Parsers
                     {
                         throw new InvalidSearchOperationException(string.Format(Core.Resources.ResourceNotSupported, targetResourceType));
                     }
+
+                    // check resource type restrictions from SMART clinical scopes
+                    if (_contextAccessor.RequestContext?.AccessControlContext?.ApplyFineGrainedAccessControl == true)
+                    {
+                        if (!ResourceTypeAllowedByClinicalScopes(targetResourceType))
+                        {
+                            throw new InvalidSearchOperationException(string.Format(Core.Resources.ResourceTypeNotAllowedRestrictedByClinicalScopes, targetResourceType));
+                        }
+                    }
                 }
             }
 
@@ -227,11 +243,34 @@ namespace Microsoft.Health.Fhir.Core.Features.Search.Expressions.Parsers
                 : searchParameter.TargetResourceTypes;
 
             ChainedExpression chainedExpression = null;
+            bool restrictedByClinicalScopes = false;
 
             foreach (var possibleTargetResourceType in possibleTargetResourceTypes)
             {
                 var wrappedTargetResourceType = new[] { possibleTargetResourceType };
                 var multipleChainType = reversed ? resourceTypes : wrappedTargetResourceType;
+
+                // check resource type restrictions from SMART clinical scopes
+                if (_contextAccessor.RequestContext?.AccessControlContext?.ApplyFineGrainedAccessControl == true)
+                {
+                    if (!ResourceTypeAllowedByClinicalScopes(possibleTargetResourceType))
+                    {
+                        restrictedByClinicalScopes = true;
+                    }
+
+                    foreach (var resourceType in multipleChainType)
+                    {
+                        if (!ResourceTypeAllowedByClinicalScopes(resourceType))
+                        {
+                            restrictedByClinicalScopes = true;
+                        }
+                    }
+
+                    if (restrictedByClinicalScopes == true)
+                    {
+                        continue;
+                    }
+                }
 
                 ChainedExpression expression;
                 try
@@ -285,6 +324,11 @@ namespace Microsoft.Health.Fhir.Core.Features.Search.Expressions.Parsers
 
             if (chainedExpression == null)
             {
+                if (restrictedByClinicalScopes)
+                {
+                    throw new InvalidSearchOperationException(Core.Resources.ChainedResourceTypesNotAllowedDueToScope);
+                }
+
                 // There was no reference that supports the search parameter.
                 throw new InvalidSearchOperationException(Core.Resources.ChainedParameterNotSupported);
             }
@@ -355,6 +399,18 @@ namespace Microsoft.Health.Fhir.Core.Features.Search.Expressions.Parsers
             {
                 input = ReadOnlySpan<char>.Empty;
             }
+        }
+
+        private bool ResourceTypeAllowedByClinicalScopes(string resourceType)
+        {
+            var resourceTypes = _contextAccessor.RequestContext?.AccessControlContext.AllowedResourceActions.Select(r => r.Resource).Distinct().ToList();
+
+            if (resourceTypes.Contains(KnownResourceTypes.All) || resourceTypes.Contains(resourceType))
+            {
+                return true;
+            }
+
+            return false;
         }
     }
 }
