@@ -114,13 +114,19 @@ namespace Microsoft.Extensions.DependencyInjection
         /// Adds background worker services.
         /// </summary>
         /// <param name="fhirServerBuilder">The FHIR server builder.</param>
+        /// <param name="addExportWorker">Whether to add the background worker for export jobs</param>
         /// <returns>The builder.</returns>
         public static IFhirServerBuilder AddBackgroundWorkers(
-            this IFhirServerBuilder fhirServerBuilder)
+            this IFhirServerBuilder fhirServerBuilder,
+            bool addExportWorker)
         {
             EnsureArg.IsNotNull(fhirServerBuilder, nameof(fhirServerBuilder));
 
-            fhirServerBuilder.Services.AddHostedService<ExportJobWorkerBackgroundService>();
+            if (addExportWorker)
+            {
+                fhirServerBuilder.Services.AddHostedService<ExportJobWorkerBackgroundService>();
+            }
+
             fhirServerBuilder.Services.AddHostedService<ReindexJobWorkerBackgroundService>();
 
             return fhirServerBuilder;
