@@ -119,6 +119,28 @@ namespace Microsoft.Health.Fhir.SqlServer.Features.Storage
             }
         }
 
+        public async Task<IReadOnlyCollection<JobInfo>> DequeueJobsAsync(byte queueType, int numberOfJobsToDequeue, string worker, int heartbeatTimeoutSec, CancellationToken cancellationToken)
+        {
+            var dequeuedJobs = new List<JobInfo>();
+
+            while (dequeuedJobs.Count < numberOfJobsToDequeue)
+            {
+                var jobInfo = await DequeueAsync(queueType, worker, heartbeatTimeoutSec, cancellationToken);
+
+                if (jobInfo != null)
+                {
+                    dequeuedJobs.Add(jobInfo);
+                }
+                else
+                {
+                    // No more jobs in queue
+                    break;
+                }
+            }
+
+            return dequeuedJobs;
+        }
+
         public async Task<JobInfo> DequeueAsync(byte queueType, string worker, int heartbeatTimeoutSec, CancellationToken cancellationToken)
         {
             try
@@ -167,7 +189,7 @@ namespace Microsoft.Health.Fhir.SqlServer.Features.Storage
                 {
                     if (sqlEx.State == 127)
                     {
-                        throw new JobConflictException(sqlEx.Message);
+                        throw new JobManagement.JobConflictException(sqlEx.Message);
                     }
 
                     throw;
